[BACKLOG-21825] Moved/Split `fields` from the `pentaho.visual.role.AbstractProperty.Type` to the internal and external subclasses, because they will have different implementations.
